performance: disable READDIRPLUS
Previously:
1. we expose the capability READDIRPLUS (which is normally a good thing)
2. therefore the kernel asks us for the READDIRPLUS result, which is a
   combined READDIR and LOOKUP (stat / GetAttr)
3. go-fuse doesn't have an API for READDIRPLUS so it calls OpenDir and
   GetAttr n times, expecting this to be the fast bit
4. we remote every call to GetAttr to the host separately, even when the
   kernel has this information in its attribute cache!

So what starts as a performance optimisation, ends up with pathological
behaviour which undermines the performance of the attribute cache.

This patch disables the READDIRPLUS capability which means that the kernel
will send its own GetAttr calls when it needs to. In the worst case this
means it will send n GetAttr calls but since these are cached they won't
be called again within 1hr (or when the path is invalidated).

Most of the times the kernel won't send any GetAttr calls as it already
has the information in the attribute cache.

Signed-off-by: David Scott <dave@recoil.org>